backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48
psycopg2-binary==2.9.10
pika==1.3.2
redis==7.3.0
python-dotenv==1.0.1
blinker==1.9.0
click==8.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
typing_extensions==4.15.0
Werkzeug==3.1.6
python-dotenv==1.0.1